refactor(frontend): group local media dialog
