+++ apps/frontend/src/routes/providers.connect.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, redirect } from "@tanstack/react-router";
 import { useState } from "react";
 import { useAuth } from "../auth";
-import { LocalVideoOpenDialog } from "../components/LocalVideoOpenDialog";
+import { LocalVideoOpenDialog } from "../components/local-media/LocalVideoOpenDialog";
 import ProviderConnectScreen from "../components/provider-connect/ProviderConnectScreen";
 import { router } from "../router";
